proc: use CPUID to determine maximum size of XSAVE area
the maximum size of the Xsave area was previously hardcoded but the
CPUID instruction can be used to determine its maximum size.
