+++ pkg/proc/amd64util/xsave_other.go
@@ -1,0 +1,7 @@
+//+build !amd64
+
+package amd64util
+
+func AMD64XstateMaxSize() int {
+	return _XSTATE_MAX_KNOWN_SIZE
+}
