+++ pkg/proc/native/ptrace_linux_386.go
@@ -9,7 +9,7 @@ import (
 
 	sys "golang.org/x/sys/unix"
 
-	"github.com/go-delve/delve/pkg/proc/linutil"
+	"github.com/go-delve/delve/pkg/proc/amd64util"
 )
 
 // ptraceGetRegset returns floating point registers of the specified thread
@@ -17,15 +17,15 @@ import (
 // See i386_linux_fetch_inferior_registers in gdb/i386-linux-nat.c.html
 // and i386_supply_xsave in gdb/i386-tdep.c.html
 // and Section 13.1 (and following) of Intel® 64 and IA-32 Architectures Software Developer’s Manual, Volume 1: Basic Architecture
-func ptraceGetRegset(tid int) (regset linutil.I386Xstate, err error) {
-	_, _, err = syscall.Syscall6(syscall.SYS_PTRACE, sys.PTRACE_GETFPREGS, uintptr(tid), uintptr(0), uintptr(unsafe.Pointer(&regset.I386PtraceFpRegs)), 0, 0)
+func ptraceGetRegset(tid int) (regset amd64util.AMD64Xstate, err error) {
+	_, _, err = syscall.Syscall6(syscall.SYS_PTRACE, sys.PTRACE_GETFPREGS, uintptr(tid), uintptr(0), uintptr(unsafe.Pointer(&regset.AMD64PtraceFpRegs)), 0, 0)
 	if err == syscall.Errno(0) || err == syscall.ENODEV {
 		// ignore ENODEV, it just means this CPU doesn't have X87 registers (??)
 		err = nil
 	}
 
-	var xstateargs [_X86_XSTATE_MAX_SIZE]byte
-	iov := sys.Iovec{Base: &xstateargs[0], Len: _X86_XSTATE_MAX_SIZE}
+	xstateargs := make([]byte, amd64util.AMD64XstateMaxSize())
+	iov := sys.Iovec{Base: &xstateargs[0], Len: uint32(len(xstateargs))}
 	_, _, err = syscall.Syscall6(syscall.SYS_PTRACE, sys.PTRACE_GETREGSET, uintptr(tid), _NT_X86_XSTATE, uintptr(unsafe.Pointer(&iov)), 0, 0)
 	if err != syscall.Errno(0) {
 		if err == syscall.ENODEV || err == syscall.EIO {
@@ -39,7 +39,7 @@ func ptraceGetRegset(tid int) (regset linutil.I386Xstate, err error) {
 	}
 
 	regset.Xsave = xstateargs[:iov.Len]
-	err = linutil.I386XstateRead(regset.Xsave, false, &regset)
+	err = amd64util.AMD64XstateRead(regset.Xsave, false, &regset)
 	return
 }
